Bug 532941 - [Wayland] Error output about "a temporary window without parent"
1) Attached snippet was broken on GTK3 with both x11 and Wayland. The
column width is 0 and the table appears as a thin line. See bug 509503.
This was fixed by always showing the table before getting preferred size
on GTK3.

2) The warnings on Wayland occurs due to having a popup window without a
parent. Shells with SWT.ON_TOP style was always created as a
GTK_WINDOW_POPUP, this causes issues when the shell has no parent on
Wayland due to lack of global coordinates. The fix is to never create a
popup window without a parent.

3) Fixed the issue with the open editor dialog (CTRL+E) having a title
bar above, by preventing decorations for shells without SHELL_TRIM style
set.

Tested with attached snippet and open editor (Ctrl + E) in child Eclipse
with GTK3.22 on X11 and Wayland.

Change-Id: I6d33c2e9bb00a55f84a79303f7326d0c8cd0a3ec
Signed-off-by: Xi Yan <xixiyan@redhat.com>