[WIP] Bug 537960 - Tree in SashForm doesnt paint last rows on resize
When resizing tree views, e.g. the Package Explorer, the text and images
of bottom rows are sometimes not painted.

In particular, JFace OwnerDrawLabelProvider registers an paint and an
item measure listener. Due to the registered measure listener, the paint
listener is not triggered when the sash is moved.

This behavior is seen due to the fix for bug 531667, which (among other
changes) restricts widgets from painting over other widgets.

This change adds an SWT bug snippet for the problem.

Change-Id: I2f523565675907aaadad8e09a0f054698272885d
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>