Bug 537960 - Tree and Table queue a re-draw if bounds changed
If an SWT tree or table has a registered listener for item measure,
paint or erase events, its possible that the last rows of the tree or
table are not correctly repainted on resize. If such a listener takes
"too long", GTK will not always invalidate the pixel cache of the
underlying tree widget.

As a result, views such as the Package Explorer and Outline will not
always draw text and icons for their last (3 to 4) rows, when the height
of their part stack is increased by dragging a Sash.

With this change, SWT trees and tables check if a resize occurred since
the last draw. If so, the current draw is discarded and a new draw is
requested via GTK API gtk_widget_queue_draw. This ensures that the GTK
pixel cache is invalidated after a sash resize, as well as minimizes
extra widget draws.

Change-Id: I84b73826ff8b566d102c5e42177576feb28a2bd6
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>