Bug 533469 - [GTK3] Ghost button since upgrade to GTK 3.22
On GTK3.19+, widget are forced to be allocated the default minimum size
regardless of the size of the fixed container, introduced after GTK
commit 3e069428. This causes widgets with 0x0 size to be visible but
unusable. The fix is to make 0x0 widgets invisible to the user
explicitly using gtk_widget_hide. Resizing widget to a non-zero size
later on shows the widget in setBounds.

Tested with snippet, empty shell when size of Control were not set. No
ghost button on Papyrus Property dialog.

Change-Id: I6e47d838c3a4679e5b8f2140601b23ca57e13ee7
Signed-off-by: Xi Yan <xixiyan@redhat.com>