Bug 536974: Orca and flat review problems
The bug here was that not all SwtFixedAccessible instances had
gtk_accessible_set_widget() called on them -- this was only done for
SwtFixedAccessible instances with a matching Java accessible.

The fix is to call gtk_accessible_set_widget() for ALL
SwtFixedAccessible instances, so that the "intermediate" accessibles in
the widget tree still function as expected.

Tested using Orca flat review mode and Accerciser.

Change-Id: I3dbcde58cc7ed744b2ec5fed6fb59d2790c88f4d
Signed-off-by: Eric Williams <ericwill@redhat.com>