Bug 535068: [Dark Theme] CTabFolder Chevron looks bad
The chevron icon had light transparency artifacts because white was
used as the one transparent color. But not all pixels are either white
(transparent) or have the icons foreground color. This "almost white"
pixels caused these ugly transparency artifacts and a very visible as
they are light in the dark theme.

The solution taken now is to use the background color or the parent
control as transparency color. This background color is not exactly the
same color as the chevron icon has but it's dark in the dark theme and
light in the light theme.
As still transparency is used there are still transparency artifacts
but they less visible as the are no longer light in the dark theme but
also dark.


Change-Id: I085dc4108ee68daa94f9eb1ae902ed0c429cfb05
Signed-off-by: Matthias Becker <ma.becker@sap.com>
Signed-off-by: Wim Jongman <wim.jongman@remainsoftware.com>