Bug 538612 - DateTime Spinner increment button trimmed
When the initial text is not the longest length possible for SWT.DATE or
SWT.TIME entry, the spinner buttons gets trimmed when the text is
changed to have longer length. Changed default date/time format
constants to the correct pattern. Implemented computeMaxTextSize method
to get the native text entry size corresponding to the longest possible
formatted text. Used the longest text size as the preferred width.

Tested on GTK3.22 with ControlExample->DateTime with SWT.DATE and
SWT.TIME on all styles (SHORT, MEDIUM, LONG). No cutoff with buttons or
warnings when incrementing values.

Change-Id: I1e62375f4af223d7ee1893d9789321a4805c5ff8
Signed-off-by: Xi Yan <xixiyan@redhat.com>