Bug 529534 - View content not fully updated after closing CTabFolder tab
Whenever an inactive tab is closed in a CTabFolder with a wrapped top
right control, the tab area is not fully updated. This can result in an
empty grey line, until the tab folder is resized.

This change adds a layout request whenever an item of the CTabFolder is
closed, ensuring correct layout of the current tab's client area. The
change also adds a snippet show-casing the SWT bug.

Change-Id: I0763b4e1a7c2e9aac08d030127e00f91e1087571
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>