Bug 464228 - [Gtk3] Combo box styling differences
Button on non-readonly Combo is not styled, so the background color of
the button on combo is dependent on OS theme. GtkEntry and
GtkToggleButton in Combo needs to be themed separately using different
css providers. Added provider instance for button to load the background
CSS to both entry and button instead
of the entry only.

Tested in child Eclipse in dark theme. On Ubuntu, Combo button is now
dark instead of white. On Fedora, Combo button now has same background
as entry.

Change-Id: I77fbe01361c84309aadc9ab187ebaf80f560850c
Signed-off-by: Xi Yan <xixiyan@redhat.com>