Bug 539367: [GTK3] Long text in Combo drawn over drop-down button and outside of combo
Combos have a GtkCellView, which renders the text, and other icons. by
default, GtkCellView has a "fit-model" property which is set to true.
This means the content in the cell view will always expand to the
maximum size. Setting it to false allows us to resize the cell view as
desired.

The fix for this bug is to set the clipping of the cell view to a
smaller size: we find the x position of the drop down icon, minus its
width. This stops the text from being drawn over the drop down icon, and
sometimes over other widgets.

Tested on GTK3.22 and 3.20, all other versions of GTK3 are unaffected.
The snippets in this patch do not show the bug, nor does the one for bug
500703. Additionally, the keys preferences page does not exhibit any
issues either. No AllNonBrowser JUnit tests fail.

Change-Id: Iabfae03c7ee5b0279b87fa7054bc28c75d2d3b7e
Signed-off-by: Eric Williams <ericwill@redhat.com>