Bug 535083 - [Wayland] setRegion() and transparent shells don't work. Platform Part-Drag causes eclipse black-out.
1) Replaced gdk_window_shape_combine_region with
gtk_widget_shape_combine_region wrapper in Control#setRegion.

2) There is a bug in GTK on Wayland where pixels in window outside
shape_region is black instead of transparent when the GtkWidget has
opacity of 1.0. This causes the problem where part-drag rectangles
introduces a black overlay shell covering Eclipse itself. This patch
uses a hack where the opacity of the widget is set to 0.99 whenever we
want to set a custom region.

3) Make the GTK_WINDOW_POPUP overlay in Tracker transient to the parent
if possible to avoid window not mapped warning on Wayland.

Tested on Wayland with Snippet134, Snippet 285, and child Eclipse.

Change-Id: I5f2a1bdd3a7acee87b09ab23f8951da0dbbd5c53
Signed-off-by: Xi Yan <xixiyan@redhat.com>