Bug 526758 - [win32] Native crash in msctf.dll::CCompositeContextAdapter::OnUnregistered()
As far as I can understand, the bug is in Windows. Specifically,
'CCompositeContextAdapter' class in 'msctf.dll'.

This class remembers pointer to 'CInputContext', which is not cleared
by 'ImmAssociateContext(0)'. After the call to 'ImmDestroyContext',
the pointer becomes invalid, and trying to use it can cause crash in
'CCompositeContextAdapter::OnUnregistered'.

The problem is related to setting focus to an edit. There's a method
'CCompositeContextAdapter::OnFocusChange' that saves and forgets the
pointer to 'CInputContext'. If the pointer is saved and 'Shell' is
closed then 'CCompositeContextAdapter::OnUnregistered' accesses
invalid pointer.

Easily reproducible with Application Verifier configured for Basics/Heaps,
because Application Verifier reduces the chances to access freed memory
to almost zero. Without it, freed memory might still be there (but it
could be overwritten by the new owner, though, causing other problems).

Code snippet that reproduces the problem:
	Display display = new Display ();
	Shell shell = new Shell(display);
	new Text(shell, SWT.BORDER);
	shell.open ();
	shell.close();
	display.dispose ();

'CCompositeContextAdapter' appeared between 10.0.16251.0 and 10.0.16288.1,
both of which are insider previews for 10.0.16299 (Fall Creators Update).
I believe this is related to the new feature of Fall Creators Update
which allows to enter emoji with Win+. hotkey. My theory is that they
had to mix two different IME editors: one that has always been there,
and the other is for emoji. Hence this new class that combines multiple
'CInputContext' together.

I evaluated three different workarounds:

1)	Calling ImmDestroyContext' strictly after all its users were destroyed.
	This sounds like the safest approach to me, so I took this path.

2)	Restoring old context in 'Control.releaseWidget'
	Debugging indicates that restoring old context causes bugged
	'CCompositeContextAdapter' to forget the soon-to-be-destroyed
	context and save the old context instead.

	However, it could be possible that old context was also destroyed
	in the mean time, for example I can imagine this scenario with two
	Shells: "Parent" and "Child".

	a) "Parent" Shell is created, allocating a new context.
	b) "Child"  Shell is created, using "Parent" as parent.
	c) "Child"  Shell inherits context from Shell "P" (in Control.createHandle)
	d) "Child"  Shell creates its own context, remembering inherited context.
	e) "Parent" Shell is closed, destroying the context.
	f) "Child"  Shell is closed, restoring inherited dead context.

3) Setting focus to something else before destroying a Shell
	Debugging indicates that setting focus to something else helps,
	because 'CCompositeContextAdapter::OnFocusChange' forgets the pointers
	when its control does not have focus.

	However, this approach can cause unwanted focus change events in
	user programs, potentially breaking them.

Change-Id: Iaba4b2e552484946b21d3d75d7cd7b3363189bc8
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>