Bug 535075 - [Wayland] application rendered with an offset relatively to host window
On Wayland, calling gtk_widget_size_allocate with a relative position to
the GtkContainer causes window contents to be rendered outside of the
actual GtkWindow. The fix is to map container's coordinate relative to
the window's coordinate, and render the position according to the
relative coordinate.

Tested with child Eclipse Tip of the Day on Wayland (Bug 536153).

Change-Id: I93d26be94f57fcb7556f4fe683f41783d5ebcdbc
Signed-off-by: Xi Yan <xixiyan@redhat.com>