Bug 540083 - [GTK3] Compare editor arrows broken on scroll
Git compare editor arrows does not render when scolling. This issue can
be seen on Wayland at all times and on X11 after deprecating
gdk_cairo_create(). This is because GTK cannot guarantee what you draw
directly on a windowing surface is going to be rendered on screen. The
fix is to queue draw events to ensure it gets rendered.

Tested in child Eclipse on X11 and Wayland.

Change-Id: I1125b324147f05f28637a824d02f95b8f203e0af
Signed-off-by: Xi Yan <xixiyan@redhat.com>