Bug 540298 - adressing initial widget size/visibility issues
This is an attempt to fix bug 533469 (and the side effect of it caused
bug 540298), by removing the workaround for setInitialBounds() coming
from GTK2 days, where SWT set the widget bounds AND visibility to the
newly created widgets with state containing ZERO_WIDTH and ZERO_HEIGHT
flags.

On GTK 3.20+ this seem to cause troubles resulting in ghost buttons
shown.

NOTE: this patch does not fix the Group widget to be shown even if the
widget size is not set (similar to bug 540298). To fix it,
Group.mustBeVisibleOnInitBounds() should return false. However, this
breaks Group size or layout, so that many groups with widgets are "too
small" to show the last widget.

Change-Id: I10d56b0acf8d063e83b3230b3e99cdd84455804b
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>