Bug 541185 - [Wayland] Help popup close after a click
1) Wayland sends an extra focus out event when clicking into the popup
after grabbing keyboard focus, which causes popup to be closed.
Workaround is to ignore this extra focus out event to fix the issue with
javadoc close after a click.

2) Current workaround that hides shell before grabbing focus is not
necessary for normal popups without a child. Deal with the case a popup
has a popup child separately in bringToTop().

This patch fixes the issue with clicking into Javadoc links, and
highlighting text in Javadoc popups. However, there is still some
remaining issues (perhaps unrelated) with auto-completion: use Tab to
switch focus to the completion details page, clicking works but
highlighting the text closes the popup. (but this issue doesn't happen
if we use mouse to switch focus to completion details page)

Change-Id: Ic095b64c2eaafdeeca65b6e61937c7d31256c97f
Signed-off-by: Xi Yan <xixiyan@redhat.com>