Bug 540163 - [Wayland] Markeplace dialog misrendered
When Marketplace dialog is opened for the first time, it somehow does
not open before Shell#forceResize is called on Wayland, and GTK gives an
incorrect allocation to the box container. The fix is to use the
calculated box size if bounds have previously been set, so that
Composite#getClientArea called from Layout#layout returns the right
area.

Change-Id: I12851f5e2f8965246e94c64cef5711895df1e5ca
Signed-off-by: Xi Yan <xixiyan@redhat.com>