Bug 541720 - [Gtk3] Regression: TableColumn.pack() clears header in 4.9
Previous patch (https://git.eclipse.org/r/#/c/126904/ )fixes table
appearing as thin line by always showing the table before getting the
preferred size, and hiding it afterwards. This caused a regression that
when the second time pack() is called, it always hides the header. This
patch fixes this by only hiding the header when it is not suppose to be
visible.

Change-Id: I7cbe06cb36a7ae7b18a52010361f5c575d88ff0b
Signed-off-by: Xi Yan <xixiyan@redhat.com>