Bug 541854 - [Wayland] Ctrl+1 popup issues
Attempting to address all popup related issue (see bug 542104).

1) Patch to bug 515773 ungrabbed focus when shell is hidden in
setVisible(false) and used
showing/hiding hack to avoid being stuck with no focus, which introduced
some of the related issues. Rewrote logic to ungrab focus after shell is
disposed.
   - Removed showing/hiding hack that uses callback to show shell in
bringToTop().
   - Addressed different cases for single popup and popups
attached to a parent popup.
   - Added additional guard to ungrab the grabbed focus..

2) Patch to bug 540166 uses callback connected to gtk destroy signal
which causes crash in Flatpak due to invalid child handle passed to the
callback. Workaround is to move the logic to SWT level by keeping track
of the child popup shell of a parent popup shell, and dispose the child
popup whenever the parent popup shell is disposed.

Known limitation:
- For the auto-completion, if you switch focus to the details page using
Tab, clicking on an item on the list page will close the auto-completion
popup instead of highlighting the list item. This is because we ungrab
the focus whenever user clicks outside the focused popup window (in this
case its the details page) to avoid being stuck without focus. This
means we can only switch between list item using the mouse when the
details page is not in focus (i.e. hit Esc first to bring focus back to
the list item before clicking list item).

- However, we should not fix this with the show/hide hack because with
the show/hide hack, hitting Esc does not work when focus is switched to
the details page using Tab.

See bug 542104 comment #2 for test cases.

Change-Id: I4ba90afdf9664e6858aa3c28cf51825f85171a76
Signed-off-by: Xi Yan <xixiyan@redhat.com>