Bug 532074 - [Wayland] Arrow menu from fast view pops up in wrong location
setLocation method is limited on Wayland since we cannot use
gdk_popup_at_rect if GdkWindow of the shell is not mapped. When the
Problems View is a fast view on Eclipse start, the drop down menu takes
PartRenderingEngine's limbo shell (see
PartRenderingEngine#safeCreateGui) as the parent shell which is off the
screen. The limbo shell's GdkWindow cannot be used on Wayland to popup
the menu at a relative location. In this case, we open the menu at the
pointer to avoid it being positioned in the wrong location.

Change-Id: Ifa120a0e68e0d012b09c1f4806dc959bb447b8f2
Signed-off-by: Xi Yan <xixiyan@redhat.com>