Bug 535978: [GTK3] Bad Scrolling behaviour of tables with Checkbox
The issue at hand in this bug is that Tree/Table editing widgets are
children of the fixedHandle, which sits above the GtkTreeView in the
widget hierarchy. This means that they are siblings of the tree/table,
and thus will continue to draw even on top of the tree/table headers.

The fix is to calculate the location and size of the header bars and
keep track of them in Composite. When editing widgets are drawn using
Composite.propagateDraw(), we check to see if the widgets would be about
to be drawn over a tree/table header. If they are, we raise/lower the
GdkWindow belonging to the editing widgets, depending on whether or not
it needs to be drawn or not.

Tested on GTK3.24 using the snippet attached to this ticket and also the
snippet from bug 531928. No AllNonBrowser JUnit tests fail, and the
update sites editor in a child Eclipse seems to be working as intended.

Change-Id: I66d90c438a3d44afdd6bf6112deef5af680cd201
Signed-off-by: Eric Williams <ericwill@redhat.com>