Bug 541635 - [Wayland] Random behaviour when moving editor tabs
1) Block in dragDetect until mouse is released or drag is detected while
using gtk_motion_event for drag detection. See Bug 515396. This was done
using extra mouseDown and dragBegun flags to keep track of the mouse or
drag state.

2) When editor tabs are being dragged, CTabFolder receives
button_pressed event a *single* motion_notify event when drag begins.
Then the Tracker rectangle is initialize and all motion_notify event and
button_released event is received by the Tracker which is not part of
Control. This causes the issue that even when mouse is released,
dragDetect still assumes the drag has not ended. So the cursor can still
move a tab even after releasing the mouse. The fix is to extend the DND
blocking logic and update that mouse was released in Tracker.

Change-Id: I71ac7edf3f2c51349426004af514037ff537c0aa
Signed-off-by: Xi Yan <xixiyan@redhat.com>