Bug 542865: [GTK4] Snippets with any children of Shell crash upon gtk_widget_get_surface_allocation()
GTK4 no longer supports "own" surfaces for widgets, so this patch
removes this logic from the GTK4 SwtFixed implementation.

Some adjustments have to made in Control, as we do
gtk_widget_get_has_surface_or_window() checks to determine the
paintHandle. In GTK4 this is broken so always use the surface of the
fixedHandle (or clientHandle in the case of Group).

Change-Id: I09c0c096407ab67ac10edfcd21a3c9927cb5f833
Signed-off-by: Eric Williams <ericwill@redhat.com>