Bug 542940 - Method setEditor of class TreeEditor raises java nullException
Calling gtk_widget_show and then gtk_widget_hide in Control#setBounds
from TreeEditor.setEditor causes display to be null. The workaround is
to force Text to be visible on initialization so that we do not need to
force show/hide in order to set bound for the TreeEditor text. See
similar fix for Button in Bug 533469.

Tested with attached snippet, Bug533469_GhostButton,
Bug497705_setBoundsAfterSetVisible.

Change-Id: I9fe8cfb63358608700246d9a6c9355b53169045d
Signed-off-by: Xi Yan <xixiyan@redhat.com>