Bug 545528 - JavaDoc for StyledTextContent#getLineAtOffset(int) is contradictory
JavaDoc defines getLineAtOffset(getLength()) should return number of
lines. This contradicts the example in same JavaDoc which claims
getLineAtOffset(4) == 2 (and the examples length is 4) but getLineCount
returns 3.

Changed JavaDoc to define getLineAtOffset(getLength()) as index of last
line since this matches implementation of DefaultContent and most
(or all?) other implementations.

Change-Id: I9cdd96177ad8c47f30fe248ba1813c3be5a48417
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>