Bug 459789 - [Mac] Window title bar should be dark when Eclipse is in dark theme
Use NSWindow.appearance to set the the title bar to dark or light.

Added a new setTheme method that will send the appAppearanceChanged
custom message and set the appearance of the window to either Dark or
Light. It'll be called from Platform UI code whenever there is a theme
change.