Bug 534923 - Cocoa: objc_msgSend is called incorrectly from os.c
C calling conventions may be different for vararg and non-vararg
functions. Change the generated code to not cast vararg functions
to function pointers with non-vararg profiles. Instead annotate
the objc_msgSend* functions in MacGenerator as "fixedargs=2" (meaning:
it's a vararg function with two fixed parameters), and change the
NativesGenerator to use this annotation to generate a cast to a
function pointer with a vararg profile with that many fixed
parameters. "fixedargs=0" or absence of the "fixedargs" annotation
means all parameters are fixed; there are no varargs. The behavior
in this case in unchanged.

The casting cannot be avoided altogether; the objc_msgSend*_stret
functions _must_ be casted.

Also use BOOL as return type for the cast of objc_msgSendSuper_bool,
not just for objc_msgSend_bool.

Verified that casting to a profile with varargs indeed makes the
compiler generate the assembly code that on x86_64 does clear
register %al, as it should per the ABI spec.[1] The old code
generated did not do this. (Verified with with cc -S using Apple's
clang and inspecting the assembler output; on OS X 10.14.2.)

Ran the new MacGenerator to regenerate the Cocoa OS.java, then ran
the new JNIGenerator to regenerate the Cocoa os.c. Re-built the
binary library. Set up an Eclipse using this rebuilt library and
ran all SWT tests. All tests passed locally.

[1] https://software.intel.com/sites/default/files/article/402129/mpx-linux64-abi.pdf

Change-Id: Ib03226ad9ecdf4bfcf789cd2b42f592b9a2b9021
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>