Bug 544140: Endless calls to Tree.rendererGetPreferredWidthProc after clicking on editable PropertySheetPage tree cell
Only re-draw the headers of a Table/Tree widget when a scroll event has
occurred. This narrows the use of gtk_widget_queue_draw() to the
specific case that needs it. It also avoids triggering native callbacks
as the enter Table/Tree does not need to be re-drawn.

Tested using the Bug535978_TableEditorHeader snippet as well as the
snippet attached to this bug report. A cursory check in the IDE yields
no regressions.

Change-Id: Ia1d4bdf9a55f83658b13cc8c64219efbd1175097
Signed-off-by: Eric Williams <ericwill@redhat.com>