Bug 541256 - Warnings: Gtk-WARNING **: Allocating size to SwtFixed
With changes for bug 533469 we started to observe new GTK+ warnings on
the standard error. In particular, widget allocation was set without
retrieving preferred widget size. To GTK+, this is a potential
programming error and so GTK+ emits a warning. For SWT, this warning is
not relevant, since SWT must control widget sizes at all times.

This change adds calls to gtk_widget_get_preferred_size(), which have
previously been used to prevent the GTK+ warnings. Since the code which
sets widget allocation now also asks the widget preferred size (which is
very often irrelevant for SWT), the GTK+ warnings are gone.

Change-Id: Ie91ccb26dd8d86646d0e30819e6a240244d1d7fe
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>