Bug 436085: setTabs method in Text widget is ignored when style includes SWT.SINGLE
Enable setTabs() to work with Text widgets of style SWT.SINGLE.
Previously this was not implemented, and SWT was just retuning for all
SWT.SINGLE Text widgets.

Additionally, the automated JUnit tests for such functionality were
limited to Win32 only -- I have now enabled them to run on Linux as
well. A reproducer snippet is also attached to this commit.

Change-Id: I259a4226475296ff2514c016590668e4d90be1c2
Signed-off-by: Eric Williams <ericwill@redhat.com>