Bug 488855 - [Win32] Browser widget under windows calls a method in the wrong thread
The problem is that Internet Explorer sometimes calls 'WebSite#MapUrlToZone' from
non-main thread. This causes errors in 'Widget#checkWidget' when 'WebSite' tries to
get to its 'IE' parent via .getParent(). In order to solve the problem I removed the
code in question; now 'IE' tells 'WebSite' what to do.

I also noticed that there are other similar pieces of code, for example in
'WebSite#ProcessUrlAction'. However, it uses '_getUrl().startsWith(IE.ABOUT_BLANK)'
this time. I found that 'WebSite#MapUrlToZone' also used this approach previously,
but it was changed due to performance reasons. Still, it's not clear whether all
instances of '_getUrl().startsWith(IE.ABOUT_BLANK)' can be changed safely to
'IE#isAboutBlank' and I don't have the time to test it, so I left it as is.

I tried to test the original problem for which 'WebSite#MapUrlToZone' was made,
but it no longer reproduces for me. Probably it was solved independently in new IE.

Change-Id: Icd27bb6481e4f6792fc844380d0da9dbdc82adff
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>