Bug 545344 - [GTK3] Read only Combo text small cutoff on Linux
A read-only SWT combo uses a GtkComboBoxText in GTK3. GtkComboBoxText
can be configured with css node combobox button.combo in a GTK3 theme.
When setting padding for this css node, text in the combo text entry
becomes cut off from the right. The larger the right and left padding,
the more text is cut off.

This change ensures that button padding is added to the combo size
computation, ensuring the entire combo text is drawn.

Change-Id: I00c650b5f7213c798eda4f995986f5a20477556f
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>