Bug 546035 - [GTK3][regression] Some icons aren't rendered properly anymore
Caused by ImageLoader#save not working properly when ImageData.data we
want to save in RGBA channel but ImageData is in RGB channel. Fix is to
iterate over each pixel and fix the byte ordering to get a good pixel
data buffer to pass to gdk_pixbuf_new.

Tested with child Eclipse on X11 and Wayland with hover icons.

Tested with ImageAnalyzer: loading any format image and saving it as PNG
should work. Load JPEG file (in RGB), add Alpha mask to it, then saving
it as PNG should work.

Change-Id: I60dd32e313276ebb0dc35a5646beb4b57691e665
Signed-off-by: Xi Yan <xixiyan@redhat.com>