Bug 546493: [GTK4] Button click coordinates are incorrect for some cases
This is a two-part fix:

1) Adjust the sendMouseDown() function in Control, as event coordinates
on GTK4 are already parent relative.

2) Table/Tree have headers built into the GtkTreeview's surface, meaning
the y-coordinates will include the height of the header. This throws off
any geometry-item querying code in Table/Tree. The fix is to adjust the
coordinates to account for the size of the header. GTK3 doesn't have
this problem because the header has its own GdkWindow which is separate
from the rest of the Table/Tree.

Tested using Snippet3 on GTK4.

Change-Id: Ia68a6edb376a823dc3beb57e4db084075e42138b
Signed-off-by: Eric Williams <ericwill@redhat.com>