Bug 546274: [GTK] vm crash when scrolling StyledText
The method scrollInPixels() doesn't seem to be needed. I've reverted the
fix from commit 64ae338b9c279c9bbcc158165aa451193a503c0a to preserve
stability on GTK3, but on GTK4 this method is never called.

Tested with the snippet attached, as well as the Canvas tab on
ControlExample (GTK3 and GTK4).

Change-Id: If907bf6c4c805c24eb021b687d5affbbc2e8bc3f
Signed-off-by: Eric Williams <ericwill@redhat.com>