Bug 547093 - [GTK] GNOME logoff hangs for 90 seconds after System.exit() on SWT.Dispose for Display
Change 1
--------
Fixed the session manager hang.

It started to happen after Bug 531634, on GNOME only.

The problem is caused by a combination of System.exit() during
'SWT.Dispose' for 'Display' in SWT application and an oversight in
GNOME session-manager [1].

Due to System.exit() SWT's 'SessionManagerDBus' doesn't have a chance to
respond with 'EndSessionResponse' and GNOME's session manager gets stuck
waiting for the reply.

Change 2
--------
Removed lock from APIs needed in shutdown hook to avoid the deadlock.
Lock is a problem because java's shutdown hooks are implemented as
separate threads, and the main thread will hold lock, because
session manager signals are dispatched in 'g_main_context_iteration'.
The lock shouldn't be used so often anyway, see Bug 546743.

Change 3
--------
Fixed incorrect behavior when QueryEndSession shows a message box
and when user closes it, QueryEndSession timeout elapsed and
session manager expects answer to EndSession instead.

[1] https://gitlab.gnome.org/GNOME/gnome-session/merge_requests/15

Change-Id: I891f2b15830caafc73ccfa6491a8f21c2bbc0160
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>