Bug 545226 - [GTK3] GC transforms don't affect GC.getClipping()
Changes done to fix bug 531667 and follow-up problems introduce a
regression. Namely, the result of GC.getClipping() is incorrect. Drawing
is correct, but the bounds returned by GC.getClipping() are not affected
by transformations set by the user on the GC object. This is not
consistent with previous behaviour and with behaviour on other platforms
(e.g. Windows 10).

This change re-adds the necessary transformations, in order to have a
consistent GC.getClipping() result. A snippet is added that showcases
the problem.

Change-Id: I33b06bb8338215f52cb02d248564b792a750067e
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>