Bug 535741 - [HiDPI] Dozens of NPEs in LogView.onMouseMove
While updating the Image for table and tree items we resize the image as
the image specified is in pixbuf instead of cairo surface. Due to this
resize when we try to retrieve the image using the pixbuf stored in the
imagelist it returns null. since the gtk store has resized pixbuf and
imagelist has a different pixbuf.

The solution is to update pixbuf array with new pixbuf and do the resize
only if the image is added in the current setImage call. Otherwise we
are resizing the pixbuf always (not desirable and not necessary)

Tested the use case mentioned in the bug and ran full test suite at 200%
and ran Manual Junit tests for tree and table

Change-Id: Iba6fbabd99a9660c79d13abd1ebdae28ac2012ba
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>