Bug 547986 - [GTK3] Table editor at first row not painted
The fix for bug 535978 introduces a regression, namely first row editor
table/tree editors are not drawn if the table/tree has a header.

This is the case, since Composite.propagateDraw will lower the GTK+
window of children widgets, if their placement is "underneath" the
table/tree header height. This is done to prevent widgets from being
drawn on top of the table/tree header when the table/tree is scrolled
down.

The check whether the widget is placed "underneath" the header is done
with the assumption that widgets in the table/tree client area have y
values (of their GTK+ allocation) starting below the header (e.g.
starting from 20+ pixels). However, those y values start with 0. As a
result, widgets of the first row are always considered to be
"underneath" the table/tree header and are so not painted.

With this change, only widgets with negative y values are considered to
be "underneath" the table/tree header. A negative y value (of the GTK+
allocation) results from scrolling down, since the widget (e.g. a
table/tree editor) starts "before" the table/tree client area.

As a result, the first row shows children widgets (such as editors) of
tables/trees. When scrolling down, widgets become not visible as soon as
part of them is covered by the table/tree header. This is not optimal
(ideally they should be partially visible), but should be enough. The
SWT mechanism to paint children (such as editors) in tables/trees does
not use native GTK mechanism but builds on top. So drawing such children
widgets partially is likely not possible or is very difficult.

Change-Id: I06e125a874895c00c40d2ccfd147c9f6ed3bf210
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>