Bug 545032: [GTK] Implement native ImageLoader
Most ImageData's created on GTK have an RGB byte order, however some
exceptions do exist. Specifically, ImageData's created from JFace's
CompositeImageDescriptor can have a BGR byte order. Because of
this, it's necessary to check the PaletteData in ImageLoader.save().

If the PaletteData indicates a negative blue shift, we are dealing with
BGR byte ordering and must adjust the offsets accordingly.

Revert some smaller changes from the original ImageLoader patch, as they
are not needed (causes the white background on breadcrumb icons).

Verified with the relevant bug snippets, JUnits, and with the JDT icons
in a child Eclipse.

Change-Id: I8d6e1cea4c1559689e42dedf334c550435cd7533
Signed-off-by: Eric Williams <ericwill@redhat.com>