Bug 546490: [GTK] Incorrect Table/Tree size
Address Table sizing issues. The majority of the issues
stem from the fact that GTK reports a 1x1 size for
Widgets which are not visible. This afflicts columns
in Trees/Tables, and throws of computeSize() values.

This patch addresses this by initially setting the column
buttons visible whenever doing a computeSize()
operation.

Further more, GtkScrolledWindows without scrollbars
sometimes do not change their size, likely due to
caching. In order to work around, we check if the Table
has an initial computed height equalling the header
height, and bump that height by 1px to trigger a resize.

Tested on GTK3.24 on X11 and Wayland. No AllNonBrowser
JUnit tests fail.

Change-Id: I35330f43a2f75b5a0f08fec47dea8a40c1b2b902
Signed-off-by: Eric Williams <ericwill@redhat.com>