Bug 541427: [GTK3][X11] Tree with no items and the size of the header is blank
A change in the GTK3.20 drawing model means that Trees without any items will
have blank headers (not drawn). The fix is to ensure that a window-system
native window is created for the Tree. Only X11 is affected by this bug,
Trees on Wayland behave normally.

Tested on GTK3.24 on X11 and Wayland, using the snippet attached and a
child Eclipse. No AllNonBrowser JUnit tests fail.

Change-Id: I5f8528b43a5ed41d652c663637673385d5eed945
Signed-off-by: Eric Williams <ericwill@redhat.com>