Bug 546248: [GTK3] Composite receives to much repaint events
As of GTK3.16, GTK will send extra draw events to SwtFixed instances
containing overlay scrollbars. This happens when the mouse leaves
an SwtFixed instance that has overlay scrollbars attached to it.

The fix is to check the dirty region in need of a redraw, and only
sent an SWT.Paint event if the actual SwtFixed instance is in need
of a redraw.

Tested on GTK3.24 on X11 and Wayland. No AllNonBrowser JUnit tests
fail.

Change-Id: I958b4c16020a14f6f24e6e3d1ca67aba2ebec329
Signed-off-by: Eric Williams <ericwill@redhat.com>