Bug 546914: [GTK] Git Reflog - branch symbol on the upper right corner does not react
Adding a control to a ToolItem via setControl() has a bug on GTK3 (all versions) where
events sent to that control are no longer processed/fired. This is because SWT-GTK3
requires a few extra steps, namely fixing the Z-order (GdkWindow stacking order)
after calling setControl(). Events in GTK3 travel via GdkWindow, so the wrong
GdkWindow on top will block events from reaching the set control.

This is visible in EGit and not SWT (Snippet58) because of the way ToolBar managers
in JFace dynamically dispose/create ToolItems on the fly.

The fix is call fixZOrder() in the parent ToolBar, everytime ToolItem.setControl()
is called. Furthermore, the map signal needs to be connected on the ToolItem's
topHandle().

Tested on GTK3.24, X11 and Wayland. Verified in a child Eclipse using EGit. No
AllNonBrowser JUnit tests fail.

Change-Id: I5bc5df8feda7482aa4180ee18d4d995c6942d463
Signed-off-by: Eric Williams <ericwill@redhat.com>