Bug 547373 - Progress view hides contents when pop-up menu is opened
The commit message is from the reported bug but the problem is a bit
wider. SWT can not process deferred layout requests if events are
processed in an external (modal) event loop.
Examples for this are: any dialog like ColorDialog or FileDialog,
DoDragDrop, TrackPopupMenu or any system provided menu and even window
move and resize operations.
The fix is to not defer layouts while an external event loop is running.

Change-Id: I91b594d8f5a8a1b2575e13b4a1fa07890e275a3c
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>