Bug 545953: Device#getDPI method can return incorrect values
Use gdk_screen_get_resolution() as width/height returned by GDK
does not take EDID into account, and is therefore inaccurate.
Furthermore, getScreenDPI() and getDPI() are the same so remove
the duplicate method and simplify the code that was calling it.

Tested on GTK3.24 on X11, no AllNonBrowser JUnit tests fail.

Change-Id: I16b22d734b3b153af0f1161bb9df3996d58b7f19
Signed-off-by: Eric Williams <ericwill@redhat.com>