Bug 393269 - [GTK] Slider Selection events do not have detail field set
Connect a custom Callback to ensure the proper detail field is set.
Previously this wasn't working because the double parameter in the
Callback's method signature was being read as a long, causing
garbage input for other parameters.

Tested on GTK3.24 and X11 using Snippet17. No adverse effects observed
in the IDE, and no AllNonBrowser JUnit tests fail.

Change-Id: I4e330cf52509011e873c197d01b0265dad861d28
Signed-off-by: Eric Williams <ericwill@redhat.com>