Bug 546552: [GTK3] Checkboxes with custom color are hard to read
Some check, radio, and toggle buttons have solid background
colors on certain GTK themes (namely the new Adwaita theme,
and the Yaru theme). Furthermore, setting a solid background
color in these cases causes the checkmark/indicator to be
nearly invisible. This patch as three parts.

Part one: use the default GTK CSS (widget name) instead
of the universal selector that is usually used for such buttons.
This ensures that the indicator/checkmark area retains it's
proper look when the default background is set.

Part two: adjust the CSS strings in setForegroundGdkRGBA() to
ensure all inner parts and styles of Buttons are styled.

Part three: define a new internal system color to match the
check/radio button foreground color. This ensures the correct
color is used for the actual check mark/radio indicator, when
the default background is set.

Tested on GTK3.24.11 and 3.24.10 (old Adwaita theme), as
well as using the Yaru theme. NoAllNonBrowser JUnit tests
fail.

Change-Id: I54be8b578b54537f4beb3bc91957420b3c5a563b
Signed-off-by: Eric Williams <ericwill@redhat.com>